renamed generate-content file
