+++ plugins/conversation-insights/src/sentiment-prompt.ts
@@ -1,5 +1,5 @@
 import { z } from '@botpress/sdk'
-import { LLMInput } from './generate-content'
+import { LLMInput } from './parse-content'
 import { Sentiment } from './sentiments'
 import * as bp from '.botpress'
 import * as sdk from '@botpress/sdk'
